requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.95.2
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
openai==2.32.0
pydantic==1.10.26
pydantic-core==2.46.3
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.39.0
